feat: append standard library
